Move url util to where it's solely used
