--- pkg/urls/join.go
@@ -1,34 +1,0 @@
-/*
-Copyright 2019 The Kubernetes Authors.
-
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
-
-    http://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.
-*/
-
-package urls
-
-import "strings"
-
-func Join(base string, others ...string) string {
-	u := base
-	for _, o := range others {
-		if !strings.HasSuffix(u, "/") {
-			u += "/"
-		}
-		if strings.HasPrefix(o, "/") {
-			u += o[1:]
-		} else {
-			u += o
-		}
-	}
-	return u
-}
